+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Show a checkbox when select overdue patients feature flag is enable
 - Handle selected overdue appointments state in overdue search screen
 - Bump kotlin to v1.7.10
+- Hide overdue patients list item checkbox by default
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes