+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Handle selected overdue appointments state in overdue search screen
 - Bump kotlin to v1.7.10
 - Hide overdue patients list item checkbox by default
+- Handle selected overdue appointments state in overdue screen
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ### Fixes